+++ public/index.php
@@ -412,6 +412,8 @@ if (\MuseDockPanel\Controllers\SetupController::needsSetup()) {
 \MuseDockPanel\Router::post('/mail/relay/domains/{id}/delete', 'MailController@relayDomainDelete');
 \MuseDockPanel\Router::post('/mail/relay/users/store', 'MailController@relayUserStore');
 \MuseDockPanel\Router::post('/mail/relay/users/{id}/delete', 'MailController@relayUserDelete');
+\MuseDockPanel\Router::post('/mail/migrations/preflight', 'MailController@migrationPreflight');
+\MuseDockPanel\Router::post('/mail/migrations/relay/execute', 'MailController@migrationRelayExecute');
 
 // Settings
 \MuseDockPanel\Router::get('/settings/services', 'SettingsController@services');